Rename files and give credits.
